+++ src/components/TeacherComponents/OnboardingForm/OnboardingForm.tsx
@@ -15,11 +15,19 @@ import {
 } from "react-hook-form";
 import { ChangeEvent } from "react";
 import { useRouter } from "next/router";
+import { useUser } from "@clerk/nextjs";
 
 import { OnboardingFormProps } from "./OnboardingForm.schema";
 
 import Logo from "@/components/AppComponents/Logo";
 import { resolveOakHref } from "@/common-lib/urls";
+import useAnalytics from "@/context/Analytics/useAnalytics";
+import useUtmParams from "@/hooks/useUtmParams";
+import getHubspotUserToken from "@/browser-lib/hubspot/forms/getHubspotUserToken";
+import getBrowserConfig from "@/browser-lib/getBrowserConfig";
+import { getHubspotOnboardingFormPayload } from "@/browser-lib/hubspot/forms/getHubspotFormPayloads";
+import { hubspotSubmitForm } from "@/browser-lib/hubspot/forms";
+import OakError from "@/errors/OakError";
 
 const OnboardingForm = ({
   showNewsletterSignUp = true,
@@ -38,6 +46,11 @@ const OnboardingForm = ({
   showTermsAndConditions?: boolean;
 }) => {
   const router = useRouter();
+  const hutk = getHubspotUserToken();
+  const utmParams = useUtmParams();
+  const { posthogDistinctId } = useAnalytics();
+  const { user } = useUser();
+
   const onFormSubmit = async (data: OnboardingFormProps) => {
     if ("worksInSchool" in data) {
       router.push(
@@ -47,8 +60,37 @@ const OnboardingForm = ({
             : "onboarding-role-selection",
         }),
       );
+    } else {
+      const hubspotFormId = getBrowserConfig("hubspotOnboardingFormId");
+      const userEmail = user?.primaryEmailAddress?.emailAddress;
+      const hubspotFormPayload = getHubspotOnboardingFormPayload({
+        hutk,
+        data: {
+          ...utmParams,
+          ...data,
+          oakUserId: posthogDistinctId,
+          email: userEmail,
+        },
+      });
+
+      try {
+        await hubspotSubmitForm({
+          hubspotFormId,
+          payload: hubspotFormPayload,
+        });
+      } catch (error) {
+        if (error instanceof OakError) {
+          reportError(error);
+        } else {
+          reportError(
+            new OakError({
+              code: "hubspot/unknown",
+              originalError: error,
+            }),
+          );
+        }
+      }
     }
-    console.log("onboarding form values: ", data);
   };
 
   return (